refactor: aT만 localStorage를 통해 관리
