+++ src/lib/types/TLocalStorageKey.ts
@@ -1,0 +1,1 @@
+export type TLocalStorageKey = 'accessToken' | 'alert';
