--- src/lib/utils/localStorage/TLocalStorageKey.ts
@@ -1,1 +1,0 @@
-export type TLocalStorageKey = 'accessToken' | 'isKid' | 'provider' | 'alert';
