+++ src/lib/utils/localStorage/getLocalStorage.ts
@@ -1,4 +1,4 @@
-import { TLocalStorageKey } from './TLocalStorageKey';
+import { TLocalStorageKey } from '../../types/TLocalStorageKey';
 
 /**
  * @param key 필요 시 key를 TKey에 추가해주세요.
