+++ src/pages/Manage/Guides.tsx
@@ -1,13 +1,14 @@
 import AppBar from '@components/layout/AppBar';
 import GuideTemplate from '@components/manage/guides/GuideTemplate';
-import getLocalStorage from '@lib/utils/localStorage/getLocalStorage';
+import { useAppSelector } from '@store/app/hooks';
+import { selectIsKid } from '@store/slices/authSlice';
 
 const Guides = () => {
-  const isKid = getLocalStorage('isKid');
+  const isKid = useAppSelector(selectIsKid);
   return (
     <>
       <AppBar label="서비스 이용 방법" to="/manage" />
-      <GuideTemplate page={'manage'} isKid={isKid} />
+      <GuideTemplate page={'manage'} isKid={isKid!} />
     </>
   );
 };
